pyproject.toml
[tool.poetry]
name = "logger-inspection"
version = "0.1.0"
description = "implementation of python logger inspection"
authors = ["yss <waku.pdxppxzy@gmail.com>"]
readme = "README.md"
packages = [{include = "logger_inspection"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
